+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,6 +1,6 @@
 package com.techcourse.controller;
 
-import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
