+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServlet.java
@@ -1,12 +1,11 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
+import com.interface21.webmvc.servlet.mvc.handler.HandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.handler.HandlerMapping;
-import com.interface21.webmvc.servlet.mvc.handler.adapter.ControllerHandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.handler.adapter.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.handler.adapter.HandlerExecutionHandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
+import com.interface21.webmvc.servlet.mvc.handler.annotation.AnnotationHandlerMapping;
+import com.interface21.webmvc.servlet.mvc.handler.annotation.HandlerExecutionHandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.handler.controller.ControllerHandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.handler.controller.ManualHandlerMapping;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
